feat(asset_platforms)_: split server.go
