+++ market-fetcher/api/handlers_utils.go
@@ -1,0 +1,54 @@
+package api
+
+import (
+	"bytes"
+	"context"
+	"crypto/md5"
+	"encoding/hex"
+	"encoding/json"
+	"log"
+	"net/http"
+	"strconv"
+	"time"
+)
+
+// sendJSONResponse is a common wrapper for JSON responses that sets Content-Type,
+// Content-Length, and ETag headers
+func (s *Server) sendJSONResponse(w http.ResponseWriter, data interface{}) {
+	// Marshal the data to calculate content length and ETag
+	buffer := &bytes.Buffer{}
+	encoder := json.NewEncoder(buffer)
+	err := encoder.Encode(data)
+	if err != nil {
+		http.Error(w, "Error encoding response", http.StatusInternalServerError)
+		return
+	}
+
+	responseBytes := buffer.Bytes()
+
+	// Calculate ETag (MD5 hash of the response)
+	hash := md5.Sum(responseBytes)
+	etag := hex.EncodeToString(hash[:])
+
+	// Set headers
+	w.Header().Set("Content-Type", "application/json")
+	w.Header().Set("Content-Length", strconv.Itoa(len(responseBytes)))
+	w.Header().Set("ETag", "\""+etag+"\"")
+
+	// Write the response
+	if _, err := w.Write(responseBytes); err != nil {
+		log.Printf("Error writing response: %v", err)
+		return
+	}
+}
+
+// Stop gracefully shuts down the server
+func (s *Server) Stop() {
+	if s.server != nil {
+		ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
+		defer cancel()
+		if err := s.server.Shutdown(ctx); err != nil {
+			log.Printf("Error shutting down server: %v", err)
+		}
+	}
+}
